# fix(core): optimize memory allocations during blob decoding