## Description

In the past, when we were using RaptorQ for the 1D encoding, it was necessary to keep around intermediate decoding state as the exact number of slivers needed for decoding was unknown and we would potentially have to reuse the decoder for a new decoding attempt.

This is no longer the case with Reed-Solomon, as we know the exact number of slivers needed for decoding. As a result, we can allocate a single 1D decoder and reuse it for all sub-decodings during blob decoding.

This commit also makes some other changes to reduce the number of allocations during blob decoding, specifically by using flat arrays instead of nested vectors where possible.

Both together result in both a **major reduction in memory usage** and a **significant performance improvement**:

- The peak memory usage while reading a 1GiB sliver from the public Testnet (before the blob re-encoding) was reduced from more than 10GiB to ~2.5GiB.
- The decoding time (without re-encoding) for 1GiB blobs was reduced by more than 60% (corresponding to a ~160% throughput improvement).

<details><summary>Benchmark details</summary>

```console
$ cargo bench -p walrus-core blob_decoding/decode/ -- --baseline main
     Running benches/basic_encoding.rs (target/release/deps/basic_encoding-669d9b131c7330c5)
     Running benches/blob_encoding.rs (target/release/deps/blob_encoding-6ae446cf9853bda3)
blob_decoding/decode/1B time:   [237.42 ms 239.74 ms 242.12 ms]
                        thrpt:  [4.1302   B/s 4.1712   B/s 4.2119   B/s]
                 change:
                        time:   [-17.007% -16.021% -14.985%] (p = 0.00 < 0.05)
                        thrpt:  [+17.626% +19.077% +20.492%]
                        Performance has improved.
blob_decoding/decode/1KiB
                        time:   [236.83 ms 238.76 ms 241.03 ms]
                        thrpt:  [4.1488 KiB/s 4.1883 KiB/s 4.2224 KiB/s]
                 change:
                        time:   [-20.303% -18.427% -16.803%] (p = 0.00 < 0.05)
                        thrpt:  [+20.196% +22.590% +25.475%]
                        Performance has improved.
blob_decoding/decode/1MiB
                        time:   [238.91 ms 240.91 ms 243.52 ms]
                        thrpt:  [4.1065 MiB/s 4.1510 MiB/s 4.1856 MiB/s]
                 change:
                        time:   [-18.806% -17.945% -16.974%] (p = 0.00 < 0.05)
                        thrpt:  [+20.444% +21.869% +23.162%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
blob_decoding/decode/16MiB
                        time:   [285.12 ms 286.72 ms 289.04 ms]
                        thrpt:  [55.356 MiB/s 55.804 MiB/s 56.117 MiB/s]
                 change:
                        time:   [-39.476% -38.440% -37.370%] (p = 0.00 < 0.05)
                        thrpt:  [+59.667% +62.443% +65.223%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Benchmarking blob_decoding/decode/256MiB: Warming up for 10.000 ms
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 11.4s.
blob_decoding/decode/256MiB
                        time:   [1.0962 s 1.1041 s 1.1144 s]
                        thrpt:  [229.72 MiB/s 231.86 MiB/s 233.54 MiB/s]
                 change:
                        time:   [-60.495% -60.093% -59.635%] (p = 0.00 < 0.05)
                        thrpt:  [+147.74% +150.58% +153.13%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Benchmarking blob_decoding/decode/1GiB: Warming up for 10.000 ms
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 40.2s.
blob_decoding/decode/1GiB
                        time:   [3.9828 s 4.0173 s 4.0607 s]
                        thrpt:  [252.18 MiB/s 254.90 MiB/s 257.10 MiB/s]
                 change:
                        time:   [-62.436% -61.965% -61.439%] (p = 0.00 < 0.05)
                        thrpt:  [+159.33% +162.91% +166.21%]
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe

```

</details> 



## Test plan

Existing test suite.

Additionally checked memory usage with `cargo instruments` and performance through benchmarks (see above).

---

## Release notes

- [x] Aggregator: Major reduction in memory usage for large blobs.
- [x] CLI: Major reduction in memory usage when reading large blobs.